requirements.txt
# Quality Management Analysis AI Agent
# Core dependencies

# Web scraping and HTTP
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
httpx>=0.25.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# AI/LLM Integration
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.5

# CLI Interface
rich>=13.0.0
typer>=0.9.0

# Environment management
python-dotenv>=1.0.0

# Financial data APIs (fallback options)
yfinance>=0.2.36

# Async support
aiohttp>=3.9.0

# PDF processing (for annual reports)
PyPDF2>=3.0.0
pdfplumber>=0.10.0
Pillow>=10.0.0

# Rate limiting
ratelimit>=2.2.1

# Web Application
streamlit>=1.30.0

# PDF Generation & Visualization
reportlab>=4.0.0
matplotlib>=3.8.0
plotly>=5.18.0
kaleido>=0.2.1
